Split instance domain block/allow import/export into separate pages/controllers
