+++ app/controllers/admin/domain_allows_controller.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
-require 'csv'
-
 class Admin::DomainAllowsController < Admin::BaseController
-  include AdminExportControllerConcern
   before_action :set_domain_allow, only: [:destroy]
-  ROWS_PROCESSING_LIMIT = 20_000
 
   def new
     authorize :domain_allow, :create?
@@ -32,26 +28,6 @@ class Admin::DomainAllowsController < Admin::BaseController
     redirect_to admin_instances_path, notice: I18n.t('admin.domain_allows.destroyed_msg')
   end
 
-  def export
-    authorize :instance, :index?
-    send_export_file
-  end
-
-  def import
-    authorize :domain_allow, :create?
-    @import = Admin::Import.new(import_params)
-    parse_import_data!(export_headers)
-
-    @data.take(ROWS_PROCESSING_LIMIT).each do |row|
-      domain = row['#domain'].strip
-      next if DomainAllow.allowed?(domain)
-
-      domain_allow = DomainAllow.new(domain: domain)
-      log_action :create, domain_allow if domain_allow.save
-    end
-    redirect_to admin_instances_path, notice: I18n.t('admin.domain_allows.created_msg')
-  end
-
   private
 
   def set_domain_allow
@@ -60,21 +36,5 @@ class Admin::DomainAllowsController < Admin::BaseController
 
   def resource_params
     params.require(:domain_allow).permit(:domain)
-  end
-
-  def export_filename
-    'domain_allows.csv'
-  end
-
-  def export_headers
-    %w(#domain)
-  end
-
-  def export_data
-    CSV.generate(headers: export_headers, write_headers: true) do |content|
-      DomainAllow.allowed_domains.each do |instance|
-        content << [instance.domain]
-      end
-    end
   end
 end
