+++ config/locales/en.yml
@@ -450,6 +450,12 @@ en:
         create: Add domain
         title: Block new e-mail domain
       title: Blocked e-mail domains
+    export_domain_allows:
+      new:
+        title: Import domain allows
+    export_domain_blocks:
+      new:
+        title: Import domain blocks
     follow_recommendations:
       description_html: "<strong>Follow recommendations help new users quickly find interesting content</strong>. When a user has not interacted with others enough to form personalized follow recommendations, these accounts are recommended instead. They are re-calculated on a daily basis from a mix of accounts with the highest recent engagements and highest local follower counts for a given language."
       language: For language
