+++ spec/controllers/admin/domain_allows_controller_spec.rb
@@ -45,31 +45,4 @@ RSpec.describe Admin::DomainAllowsController, type: :controller do
       expect(response).to redirect_to(admin_instances_path)
     end
   end
-
-  describe 'GET #export' do
-    it 'renders instances' do
-      Fabricate(:domain_allow, domain: 'good.domain')
-      Fabricate(:domain_allow, domain: 'better.domain')
-
-      get :export, params: { format: :csv }
-      expect(response).to have_http_status(200)
-      expect(response.body).to eq(IO.read(File.join(self.class.fixture_path, 'files/domain_allows.csv')))
-    end
-  end
-
-  describe 'POST #import' do
-    it 'allows imported domains' do
-      post :import, params: { admin_import: { data: fixture_file_upload('files/domain_allows.csv') } }
-
-      expect(response).to redirect_to(admin_instances_path)
-
-      # Header should not be imported
-      expect(DomainAllow.where(domain: '#domain').present?).to eq(false)
-
-      # Domains should now be added
-      get :export, params: { format: :csv }
-      expect(response).to have_http_status(200)
-      expect(response.body).to eq(IO.read(File.join(self.class.fixture_path, 'files/domain_allows.csv')))
-    end
-  end
 end
